V3.0
Major changes are introduction of the 'FactorioAnalyticsAPI'. This has the ability to use the endpoint that I run locally for trial execution.

So, there is now a way to run trials without needing a copy of Factorio, or even needing an install! You can just use the public API - which may have some limits in the future depending on load.